move most of the definitions in main.py to events.py
